\name{seqGen-package}
\alias{seqGen-package}
\alias{seqGen}
\docType{package}
\title{
\packageTitle{seqGen}
}
\description{
\packageDescription{seqGen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{seqGen}
\packageIndices{seqGen}
Use \code{generateRandBioString()} to generate a random DNA string for a specified length with introduced shifts. Output is in IUPAC notation.
}
\author{
\packageAuthor{seqGen}

Maintainer: \packageMaintainer{seqGen}
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
generateRandBioString(150)
}
